Laying the infrastructure for the Stopper User Interface
Wrote the idgeter interface type and added it to the posters and geters in order to get the id for each of them.

Added the GetId function for each of the services: localfileservice, timeoutservice, multicastservice, leveldbservice, kademliaservice, googledriveservice, appsscriptservice.

Changed the getters and posters to maps from strings(their IDs) to a geter or a poster.

Modified the Register functions for each of the getters and posters to reflect the change to maps

Wrote the DeRegister functions for each of the getters and posters to dergeister them whenever the function is called.

Removed the contentservices type and Register contentservices function so that the start function can do all the individual registering and deregistgering

Created Start functions each of the services.
Moved the Register geter and posters to their respective start functions in each of the services.

Made the stop function in fuse to unmount ccfs

moved crawler starting behind the *serve flag (web).
moved ParseFlagsAndTakeAction function to Start.go
moved the block that registers the services for the object module from ccfs.go to start.go
added an "Instance" variable of type FS
added a global "Running" variable to track fuse.