More infrastructre Laying for the Stopper user interface
Wrote the stop functions for each of the services.
Updated the stop.go file to reflect the "services".stop functions written.
Changed the init function in the multi-cast, googledrive, appsscript and kademliadht services to a Start function.
Added the multicast, googledrive, appsscript and kademliadht.Start, multicast, googledrive, appsscript and kademliadht.Stop to the start and stop files.
Added the Registergetters to the multicast, googledrive and appsscript.Start function
Added the Registergetters and posters to the kademliadht.Start function