Fixed the use of Ctrl+c in stopping the command Line interface
Made continueCLI, the command line variable in ccfs.go a global variable and set it to false in the stop function so that the command line can stop whenever Stop is called.
In addition to continueCLI, added a string channel to receive the stopping command (ctrl+c) and stops the whole program when the command is issued.