Fixed some parts of googledriveservice and multicast service
Put the check for instance.conn of multicastservice in the start function so that when there isn't a network connection, multicast service doesn't panic but it returns a message that netowrk connection was not established.

Created an init function in googledriveservice_test to change the working directory to the directory of the bin file and also call Start()
We learnt that the auth code is only valid for one use and when it is gotten, if it hasn't been used, it will be valid for only 1 hour.