Modified the way the authCode is received in the googledriveservice
Changed getting code from the user from having them saving it as a txt file to having them paste it into the command line since it is a one time code.
Created the variable (runnning) and functions (Running and CommandLine) that returns the status and handles the commanLine stuff for each of the services (Have not implemented it yet).