Moved repl.go from Main to Services and Created a runner function
Wrote the timeout command line function that takes in a time and changes the timeout to the input time.
Made repl.go a file under the services package (previously in main)
Moved the status function to repl.go from Start.go in main.
Wrote init function for each of the services that registers them so they can give command to the command-line interface
Split the register/deregistercommander function into a register/deregistercommander and register/deregisterrunner functions.